More test updates: done with execution + local dag suite

+++ core/src/test/scala/org/apache/spark/monotasks/SimpleMonotask.scala
@@ -19,5 +19,5 @@ package org.apache.spark.monotasks
 import org.apache.spark.TaskContext
 
 /** A minimal subclass of monotask for use in testing. */
-class SimpleMonotask(taskAttemptId: Int)
+class SimpleMonotask(taskAttemptId: Long)
   extends Monotask(new TaskContext(null, null, 0, null, taskAttemptId))
